Move DirectWeightedSetBlueprint to separate file(s).
